🚚 Move discussion into features/poll (RAL-1291)
- Move components/discussion/discussion.tsx to
  features/poll/components/discussion.tsx
- Drop the index.ts barrel; importers reference the file directly
- Remove the freed entry from the biome.json migration allowlist
