--- apps/web/src/components/discussion/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./discussion";
