+++ apps/web/src/features/poll/components/discussion.tsx
@@ -25,13 +25,13 @@ import { OptimizedAvatarImage } from "@/components/optimized-avatar-image";
 import { Participant, ParticipantName } from "@/components/participant";
 import { useParticipants } from "@/components/participants-provider";
 import { useEditToken } from "@/components/poll/mutations";
+import TruncatedLinkify from "@/components/poll/truncated-linkify";
+import { useUser } from "@/components/user-provider";
 import { usePoll, useRole } from "@/features/poll/client";
 import { Trans, useTranslation } from "@/i18n/client";
 import { RelativeTime } from "@/lib/datetime/relative-time";
 import { requiredString } from "@/lib/utils/form-validation";
 import { trpc } from "@/trpc/client";
-import TruncatedLinkify from "../poll/truncated-linkify";
-import { useUser } from "../user-provider";
 
 interface CommentForm {
   authorName: string;
